docker-apps/edge/forecaster/app/requirements.txt
setuptools
minio
optuna
paho-mqtt
torch
numpy
scikit-learn
scipy
pandas
requests
paho-mqtt